requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
PyJWT
passlib[bcrypt]
pydantic
python-multipart
